feat: Lottos 클래스 생성
발행할 로또의 숫자를 구하는 getLottoAmount
발행할 숫자 만큼 로또를 발행하는 getLottos 메소드 추가
