--- src/Lottos/splitLottoNumber.js
@@ -1,3 +1,0 @@
-export const splitLottoNumber = (lottoString) => {
-  return lottoString.split(',').map(Number);
-};
